WIP: NPM Package
- Use webpack and babel to create an es6 / browser compatible version of the domain entities, enums, and utilities

* Added domain/index.js,
  - imports then exports all entities, enums, and utils

* In all domain entities and enums
  - remove NODE and window detection, just use module.exports and the class
  - in constructors where args are set to properties, only use nameToId if the argument was supplied, to avoid error

* In all tests and scripts,
  - get entities and name utils from the domain/index.js file

* Added dist/
  - fismo.js,
  - test.html
    - loads fismo.js and instantiates entities to test constructors

* Added webpack.config.js,
  - enters at ./scripts/domain/index.js
  - generates dist/fismo.js, ready for use in browser

* In package.json,
  - in scripts
    - changed build to build:contracts
    - added build:sdk
  - in devDependencies,
    - added @babel/core, @babel/preset-env, babel-loader, webpack, and webpack-cli
