+++ package.json
@@ -19,7 +19,8 @@
     "solidity"
   ],
   "scripts": {
-    "build": "npx hardhat compile",
+    "build:contracts": "npx hardhat compile",
+    "build:sdk": "webpack --mode development",
     "test": "npx hardhat test",
     "lint": "solhint 'contracts/**/*.sol'",
     "size": "npx hardhat size-contracts",
@@ -33,9 +34,12 @@
     "@openzeppelin/contracts": "^4.4.2"
   },
   "devDependencies": {
+    "@babel/core": "^7.17.8",
+    "@babel/preset-env": "^7.16.11",
     "@nomiclabs/hardhat-ethers": "^2.0.4",
     "@nomiclabs/hardhat-etherscan": "^3.0.0",
     "@nomiclabs/hardhat-waffle": "^2.0.2",
+    "babel-loader": "^8.2.4",
     "chai": "^4.2.0",
     "eip55": "^2.1.0",
     "ethereum-waffle": "^3.4.0",
@@ -43,7 +47,9 @@
     "hardhat": "^2.8.3",
     "hardhat-contract-sizer": "^2.4.0",
     "hardhat-gas-reporter": "^1.0.7",
+    "solhint": "^3.3.6",
     "solidity-coverage": "^0.7.18",
-    "solhint": "^3.3.6"
+    "webpack": "^4.46.0",
+    "webpack-cli": "^4.9.2"
   }
 }
