--- scripts/domain/Guard.js
@@ -1,21 +1,0 @@
-/**
- * Fismo Domain Enum: Guard
- * @author Cliff Hall <cliff@futurescale.com>
- */
-const NODE = (typeof module !== 'undefined' && typeof module.exports !== 'undefined');
-class Guard {}
-
-Guard.ENTER = 0;
-Guard.EXIT = 1;
-
-Guard.Modes = [Guard.ENTER, Guard.EXIT];
-
-// Export
-if (NODE) {
-    module.exports = Guard;
-} else {
-    if (window) {
-        if (!window.Fismo) window.Fismo = {};
-        window.Fismo.Guard = Guard;
-    }
-}
