+++ scripts/domain/index.js
@@ -1,0 +1,25 @@
+const ActionResponse = require("./entity/ActionResponse");
+const Guard = require("./enum/Guard");
+const Machine = require("./entity/Machine");
+const Position = require("./entity/Position");
+const State = require("./entity/State");
+const Transition = require("./entity/Transition");
+const {
+    nameToId,
+    validateId,
+    validateNameLax,
+    validateNameStrict
+} = require("./util/name-utils");
+
+module.exports = {
+    ActionResponse,
+    Guard,
+    Machine,
+    Position,
+    State,
+    Transition,
+    nameToId,
+    validateId,
+    validateNameLax,
+    validateNameStrict
+};
